feat: added tech record generators for each vehicle type, and merged the updated methods for creating and posting tech records into their original class
